Adding restaurant table.  Adding request table.

+++ db/migrate/20140423145549_create_restaurants.rb
@@ -8,19 +8,19 @@ class CreateRestaurants < ActiveRecord::Migration
       t.string :url
       t.string :rating_img
       t.string :snippet_img
-      t.double :rating
+      t.float :rating
       t.string :categories
       t.integer :review_count
       t.string :open_start
       t.string :open_end
       t.boolean :open
-      t.double :sum_price_scores
-      t.double :sum_food_scores
-      t.double :distance
-      t.double :distance_score
+      t.float :sum_price_scores
+      t.float :sum_food_scores
+      t.float :distance
+      t.float :distance_score
       t.text :votes
-      t.double :rating_score
-      t.double :percent_match
+      t.float :rating_score
+      t.float :percent_match
 
       t.timestamps
     end
